requirements.txt
aiosqlite==0.16.1
asgiref==3.4.1
click==8.0.1
fastapi==0.68.0
h11==0.12.0
iso8601==0.1.16
pydantic==1.8.2
pypika-tortoise==0.1.1
python-dotenv==0.19.0
pytz==2021.1
starlette==0.14.2
tortoise-orm==0.17.6
typing-extensions==3.10.0.0
uvicorn==0.14.0
passlib==1.7.4
bcrypt==3.2.0
